Benchmark comparison 

```
cpu: Intel(R) Core(TM) i7-9750H CPU @ 2.60GHz
BenchmarkMsgPackShamaton
BenchmarkMsgPackShamaton-12             	  302420	      3625 ns/op
BenchmarkMsgPackVmihailenco
BenchmarkMsgPackVmihailenco-12          	  938449	      1123 ns/op
BenchmarkProtobuf
BenchmarkProtobuf-12                    	 1583583	       693.7 ns/op
BenchmarkGOB
BenchmarkGOB-12                         	  306645	      3736 ns/op
BenchmarkDecodeMsgPackShamaton
BenchmarkDecodeMsgPackShamaton-12       	    1995	    504313 ns/op
BenchmarkDecodeMsgPackVmihailenco
BenchmarkDecodeMsgPackVmihailenco-12    	 3663260	       292.1 ns/op
BenchmarkDecodeProtobuf
BenchmarkDecodeProtobuf-12              	 1576057	       718.7 ns/op
BenchmarkDecodeGOB
BenchmarkDecodeGOB-12                   	 3314694	       334.6 ns/op
```